+++ CHANGES.rst
@@ -20,7 +20,7 @@ Changes
 
 Enhancements
 ------------
-- mocks: warn about small theta and large mu in float32 precision [#299]
+- Warn about loss of precision for float32 calculations involving small ``theta`` in ``DDtheta_mocks`` and large ``mu`` in ``DDsmu_mocks`` [#299]
 
 2.5.0 (2022-12-23)
 ================